requirements.txt
Flask
Werkzeug
pillow-heif
Pillow
easyocr
pytesseract
pandas
matplotlib
numpy
opencv-python
ultralytics
imutils
requests